@@ -1,10 +1,8 @@
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { Http, HttpModule } from '@angular/http';
+import {ModuleWithProviders, NgModule} from '@angular/core';
+import { HttpModule } from '@angular/http';
 import { RouterModule, Routes } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { LocalStorageModule } from 'angular-2-local-storage';
-
 import { AppComponent } from './app.component';
 import { MovieComponent } from './component/movie/movie.component';
 import { MovieListComponent } from './component/movie-list/movie-list.component';
@@ -19,15 +17,54 @@ import { NgxCarouselModule } from 'ngx-carousel';
 import { CategoryService } from './service/category.service';
 import { CartService } from './service/cart.service';
 import { CreateUserComponent } from './component/create-user/create-user.component';
-import 'hammerjs';
 import { UserService } from './service/user.service';
-import { HttpClientModule } from '@angular/common/http';
+import { ShowErrorsComponent } from './component/show-errors/show-errors.component';
+import { FormsHelperService } from "./service/forms-helper.service";
+import { HttpClientModule } from "@angular/common/http";
+import { StoreButtonComponentComponent } from './component/store-button-component/store-button-component.component';
+import { CartComponent } from './component/cart/cart.component';
+import { CartDetailComponent } from './component/cart-detail/cart-detail.component';
+import { LoginComponent } from './component/login/login.component';
+import {AuthenticateService} from "./service/authenticate.service";
+import { UserAccountComponent } from './component/user-account/user-account.component';
+import 'hammerjs';
 
-export const appRoutes: Routes = [
-  { path: '', component: HomePageContentComponent },
-  { path: 'movies', component: SearchResultContentComponent },
-  { path: 'movies/:id', component: MovieDetailComponent },
-  { path: 'user/create', component: CreateUserComponent }
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: HomePageContentComponent,
+  },
+  {
+    path: 'category/movie/:category',
+    component: MovieListComponent,
+    data: {
+      breadcrumb: "Liste de films"
+    },
+  },
+  {
+    path: 'movies/:id',
+    component: MovieDetailComponent,
+    data: {
+      breadcrumb: "Détail"
+    }
+  },
+  { path: 'cart', component: CartComponent, data: {
+      breadcrumb: "Panier"
+    }
+  },
+  { path: 'movies', component: SearchResultContentComponent, data: {
+      breadcrumb: "Recherche"
+    }
+  },
+  { path: 'user/create', component: CreateUserComponent, data: {
+      breadcrumb: "Création utilisateur"
+    }
+  },
+  { path: 'account/:id', component: UserAccountComponent, data: {
+      breadcrumb: "Compte"
+    }
+  }
 ];
 
 @NgModule({
@@ -41,25 +78,30 @@ export const appRoutes: Routes = [
     BreadcrumbComponent,
     SearchResultContentComponent,
     MovieHorizontalScrollComponent,
-    CreateUserComponent
+    CreateUserComponent,
+    ShowErrorsComponent,
+    StoreButtonComponentComponent,
+    CartComponent,
+    CartDetailComponent,
+    LoginComponent,
+    UserAccountComponent
   ],
   imports: [
-    LocalStorageModule.withConfig({
-      prefix: 'movie-market',
-      storageType: 'localStorage'
-    }),
     BrowserModule,
     HttpModule,
     HttpClientModule,
-    RouterModule.forRoot(appRoutes),
+    RouterModule.forRoot(routes),
     FormsModule,
+    ReactiveFormsModule,
     NgxCarouselModule
   ],
   providers: [
     MovieService,
     CategoryService,
     UserService,
-    CartService
+    CartService,
+    FormsHelperService,
+    AuthenticateService
   ],
   bootstrap: [AppComponent]
 })
