@@ -1,4 +1,12 @@
 import { Component, OnInit } from '@angular/core';
+import {ActivatedRoute, NavigationEnd, Params, PRIMARY_OUTLET, Router} from "@angular/router";
+import "rxjs/add/operator/filter";
+
+interface IBreadcrumb {
+  label: string;
+  params?: Params;
+  url: string;
+}
 
 @Component({
   selector: 'app-breadcrumb',
@@ -7,20 +15,65 @@ import { Component, OnInit } from '@angular/core';
 })
 export class BreadcrumbComponent implements OnInit {
 
-  breadcrumbs: string[];
+  breadcrumbs: IBreadcrumb[];
 
-  constructor() {
-    this.breadcrumbs = [
-      'mon',
-      'fil',
-      'd\'ariane',
-      'est',
-      'non',
-      'fonctionnel'
-    ];
-   }
-
-  ngOnInit() {
+  constructor(
+    private activatedRoute: ActivatedRoute,
+    private router: Router
+  ) {
+    this.breadcrumbs = [];
   }
 
+  ngOnInit() {
+    const ROUTE_DATA_BREADCRUMB: string = "breadcrumb";
+
+    //subscribe to the NavigationEnd event
+    this.router.events.filter(event => event instanceof NavigationEnd).subscribe(event => {
+      //set breadcrumbs
+      let root: ActivatedRoute = this.activatedRoute.root;
+      this.breadcrumbs = this.getBreadcrumbs(root);
+    });
+  }
+
+  private getBreadcrumbs(route: ActivatedRoute, url: string="", breadcrumbs: IBreadcrumb[]=[]): IBreadcrumb[] {
+    const ROUTE_DATA_BREADCRUMB: string = "breadcrumb";
+
+    //get the child routes
+    let children: ActivatedRoute[] = route.children;
+
+    //return if there are no more children
+    if (children.length === 0) {
+      return breadcrumbs;
+    }
+
+    //iterate over each children
+    for (let child of children) {
+      //verify primary route
+      if (child.outlet !== PRIMARY_OUTLET) {
+        continue;
+      }
+
+      //verify the custom data property "breadcrumb" is specified on the route
+      if (!child.snapshot.data.hasOwnProperty(ROUTE_DATA_BREADCRUMB)) {
+        return this.getBreadcrumbs(child, url, breadcrumbs);
+      }
+
+      //get the route's URL segment
+      let routeURL: string = child.snapshot.url.map(segment => segment.path).join("/");
+
+      //append route URL to URL
+      url += `/${routeURL}`;
+
+      //add breadcrumb
+      let breadcrumb: IBreadcrumb = {
+        label: child.snapshot.data[ROUTE_DATA_BREADCRUMB],
+        params: child.snapshot.params,
+        url: url
+      };
+      breadcrumbs.push(breadcrumb);
+
+      //recursive
+      return this.getBreadcrumbs(child, url, breadcrumbs);
+    }
+  }
 }
