@@ -1,4 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { CategoryService } from '../../service/category.service';
+import { Category } from '../../model/category';
+import { CartService } from './../../service/cart.service';
+import {AuthenticateService} from "../../service/authenticate.service";
 
 @Component({
   selector: 'app-top-nav-bar',
@@ -6,14 +10,20 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./top-nav-bar.component.css']
 })
 export class TopNavBarComponent implements OnInit {
+  categories: Category[];
+  error = null;
+  finished = false;
 
-  title;
-
-  constructor() {
-    this.title = 'Movie Market';
-  }
+  constructor(private categoryService: CategoryService,
+      public cartService: CartService,
+      public authenticateService: AuthenticateService) { }
 
   ngOnInit() {
+    this.categoryService.all().subscribe(
+      value => this.categories = value,
+      error => this.error = 'movieService.all error',
+      () => this.finished = true
+  );
   }
 
 }
