@@ -1,6 +1,12 @@
+import { Category } from './category';
+
 export class Movie {
-  id: number;
-  name: string;
-  img: string;
-  description: string;
+    id: number;
+    categories: Category[];
+    img: string;
+    name: string;
+    description: string;
+    price: number;
+    quantity: number;
+    releaseYear: number;
 }
