@@ -1,6 +1,8 @@
 package com.m2gi.movieMarket.domain.repository.person;
 
 import com.m2gi.movieMarket.domain.entity.person.User;
+import com.m2gi.movieMarket.domain.entity.person.UserRole;
+import org.hibernate.Hibernate;
 
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
@@ -14,7 +16,12 @@ public class UserFacade implements UserFacadeLocal {
 	private EntityManager em;
 
 	public int create(User user) {
+		UserRole userRole = new UserRole("ROLE_USER");
+
+		user.addUserRole(userRole);
+
 		this.em.persist(user);
+		this.em.flush();
 
 		return user.getId();
 	}
@@ -28,9 +35,27 @@ public class UserFacade implements UserFacadeLocal {
 	}
 
 	public User find(Object id) {
-		return this.em.find(User.class, id);
+
+		User user = this.em.find(User.class, id);
+
+		Hibernate.initialize(user.getUserRoles());
+		Hibernate.initialize(user.getAddresses());
+
+		return user;
 	}
 
+	public User findByEmail(String email) {
+		User user = (User) this.em.createQuery("select u from User as u where u.email = :email")
+				.setParameter("email", email)
+				.getSingleResult();
+
+		Hibernate.initialize(user.getUserRoles());
+		Hibernate.initialize(user.getAddresses());
+
+		return user;
+
+	}
+	
 	public List<User> findAll(int from, int to) {
 		return this.em.createQuery("select u from User as u")
 				.setFirstResult(from)
